Add physical page allocator
To get heap memory rolling we'll need a page allocator. I've opted to
implement a free list allocator like xv6 uses. I'm not particularly
happy with this implementation, Rust seems to be not super comfortable
with the safety of this allocator and I'd rather be leaning more on the
type system. But I'm happy/surprised it's working at all.

I've borrowed a bit of the linker magic from
http://osblog.stephenmarz.com
to find the extent of the heap, which seems like a nice way to do it. I
think I'll revisit this later once the whole paging implementation is
complete. With that overview I reckon I'll be able to design a better
data flow.