Add an initial page table implementation
We need to get out of machine mode and into supervisor mode, which will
enable the MMU and give us a virtual address space. To get there we'll
need to setup and manage the page table used to create the virtual
mapping. This is a (pretty ugly) first implementation of that.

For some reason pointers in rust are confusing me (I'm probably
overthinking it...) I was going to try and present clean interfaces
everywhere, but to get it working I ended up just passing pointers
around. Although I'm reasonably confident this is working, the last test
crashes the computer because it calls `mret` without setting a valid
return pointer. Oops. I'll fix that next.