Fix page table implementation and activate MMU
Alright, so I had a number of problems getting to supervisor mode and
activating the MMU. But this seems to work, even if I've made a bit of
a mess in the process.

So there were some problems in the way I was writing the page tables. It
looks like they were mostly caused by unintentionally getting addresses
of pointers, when I just wanted the pointer itself. This was happening
because I wasn't noticing when I had a borrow value rather than an owned
value. Which makes a lot of sense in hindsight, but Rust does a nice job
of hiding these details away. I know to look out for that now.

I also had trouble with the actual entry into supervisor mode. It turns
out that we have to also give physical memory permissions to supervisor
mode from machine mode. I noticed this happening in xv6, but it took me
a while to see where the RISC V spec describes this.

I don't like the `_start` routine. I think I'll change it up soonish.
But it's doing the job I need for now so it can stay.